Extract separate ABI module
